+++ src/vue/modules/create-opportunity/index.vue
@@ -297,10 +297,7 @@
           <div class="app__form-field">
             <select-field
               v-model="form.saleInformation.assetType"
-              :is-value-translatable="true"
               name="asset-create-asset-type"
-              key-as-value-text="label"
-              :values="assetTypes"
               :label="
                 'create-opportunity.investor-requirements' | globalize
               "
@@ -309,7 +306,15 @@
               :error-message="getFieldErrorMessage(
                 'form.saleInformation.assetType',
               )"
-            />
+            >
+              <option
+                v-for="assetType in assetTypes"
+                :key="assetType.value"
+                :value="assetType.value"
+              >
+                {{ assetType.label | globalize }}
+              </option>
+            </select-field>
           </div>
         </div>
         <div class="app__form-row">
@@ -783,7 +788,7 @@ export default {
       const operation = {
         requestID: requestId,
         code: this.form.information.code,
-        assetType: this.form.saleInformation.assetType.value,
+        assetType: this.form.saleInformation.assetType,
         preissuedAssetSigner: this.accountId,
         trailingDigitsCount: this.decimalPints,
         initialPreissuedAmount: this.form.information.maxIssuanceAmount,