+++ CHANGELOG.md
@@ -25,6 +25,7 @@ for further information about branching and tagging conventions.
 - `AmountInputField`
 - Cap asset field to the create sale form
 - Display amount + fee in movements list
+- "custom-select" package for displaying customized select field
 
 #### Changed
 - Now loading account converted balances in
@@ -50,6 +51,7 @@ for further information about branching and tagging conventions.
   verification form
 - Now using "d3Format" and SI prefixes while formatting chart percent hints
 - Now displaying asset for incoming withdrawal requests
+- Now passing values to select field using a slot and "option" tags
 
 #### Removed
 - Cursor pointer on disabled select